issue #1 modificando os componentes addtask e todo list para que o array seja gerenciado pelo todolist
