+++ src/app/add-task-component/add-task-component.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, Output } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { Tarefa } from '../todo-list-component/tarefa.interface';
 
 @Component({
@@ -6,18 +6,8 @@ import { Tarefa } from '../todo-list-component/tarefa.interface';
   templateUrl: './add-task-component.component.html',
   styleUrls: ['./add-task-component.component.scss']
 })
+
 export class AddTaskComponentComponent{
   
   @Output() aoAdicionar = new EventEmitter<any>();
-
-  descricao: string;
-  tarefas: Tarefa[] = [];
-  id = 0;
-  
-  adicionar(){
-    this.id +=1;
-    const tarefaNova = { id: this.id, descricao: this.descricao };
-    this.tarefas.push(tarefaNova);
-    this.aoAdicionar.emit(tarefaNova);
-  }
 }
