+++ src/app/todo-list-component/todo-list-component.component.html
@@ -4,7 +4,17 @@
     
 </div>
 <div class="div_list">
-    <app-add-task-component>(aoAdicionar)="adicionar($event)"</app-add-task-component>
+    <app-add-task-component (aoAdicionar)="adicionar()">
+    </app-add-task-component>
+    <div class="divAddTask">
+        <div>
+            <form (ngSubmit)="adicionar()">
+                <input class="inputTask" type="text" name="tarefa" id="tarefa" [(ngModel)]="descricao" placeholder="Insira sua tarefa aqui!">
+                <button id="addButton" class="botao" type="submit">Adicionar</button>
+            </form>
+            <app-list-task-component [tarefas]="tarefas"></app-list-task-component>
+        </div>  
+    </div>
 </div>
