+++ src/app/todo-list-component/todo-list-component.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from '@angular/core';
 import { AppComponent } from '../app.component';
-import { Tarefa } from './tarefa.interface';
+import { Component, EventEmitter, Input, Output, OnInit } from '@angular/core';
+import { Tarefa } from '../todo-list-component/tarefa.interface';
 
 @Component({
   selector: 'app-todo-list-component',
@@ -8,11 +8,25 @@ import { Tarefa } from './tarefa.interface';
   styleUrls: ['./todo-list-component.component.scss'],
 })
 
-export class TodoListComponentComponent implements OnInit {
+export class TodoListComponentComponent{
 
   tarefas: Tarefa[] = [];
 
   constructor() { }
+
+  @Output() aoAdicionar = new EventEmitter<any>();
+
+  descricao: string;
+  id = 0;
+  
+  adicionar(){
+    this.id +=1;
+    const tarefaNova = { id: this.id, descricao: this.descricao };
+    this.tarefas.push(tarefaNova);
+    this.aoAdicionar.emit(tarefaNova);
+  }
+
+
 
   ngOnInit(): void {
   }
